# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/recover": {
      "post": {
        "operationId": "RecoverController_recoverPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecoverPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Recover Routes"
        ]
      }
    },
    "/auth/recover/confirm": {
      "post": {
        "operationId": "RecoverController_confirmToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Recover Routes"
        ]
      }
    },
    "/auth/recover/changePassword": {
      "put": {
        "operationId": "RecoverController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Recover Routes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sex-adapt",
    "description": "API Routes and data",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EstablishmentEntity": {
        "type": "object",
        "properties": {
          "accessibilities": {
            "$ref": "#/components/schemas/AccessibilityEntity"
          },
          "favoritedBy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserEntity"
            }
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "address": {
            "type": "object",
            "properties": {
              "street": {
                "required": true,
                "type": "string"
              },
              "number": {
                "required": true,
                "type": "string"
              },
              "complement": {
                "required": true,
                "type": "string"
              },
              "cep": {
                "required": true,
                "type": "string"
              }
            }
          },
          "ground_floor_room": {
            "type": "boolean"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          },
          "cover_photo": {
            "type": "string"
          },
          "room_photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "landline": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewEntity"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "accessibilities",
          "favoritedBy",
          "name",
          "price",
          "category",
          "address",
          "ground_floor_room",
          "latitude",
          "longitude",
          "cover_photo",
          "landline",
          "reviews",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReviewEntity": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "grade": {
            "type": "number"
          },
          "comment": {
            "type": "string"
          },
          "establishment": {
            "$ref": "#/components/schemas/EstablishmentEntity"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "grade",
          "comment",
          "establishment",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "SuportEntity": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "message": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "message",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "SuggestEstablishmentEntity": {
        "type": "object",
        "properties": {
          "suggestedBy": {
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "landline": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "isInPlace": {
            "type": "boolean"
          },
          "ground_floor_room": {
            "type": "boolean"
          },
          "elevator": {
            "type": "boolean"
          },
          "bar": {
            "type": "boolean"
          },
          "unevenness": {
            "type": "boolean"
          },
          "incompatible_dimensions": {
            "type": "boolean"
          },
          "sign_language": {
            "type": "boolean"
          },
          "tactile_floor": {
            "type": "boolean"
          },
          "braille": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "suggestedBy",
          "name",
          "category",
          "address",
          "landline",
          "price",
          "website",
          "isInPlace",
          "ground_floor_room",
          "elevator",
          "bar",
          "unevenness",
          "incompatible_dimensions",
          "sign_language",
          "tactile_floor",
          "braille",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "accessibilities": {
            "$ref": "#/components/schemas/AccessibilityEntity"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewEntity"
            }
          },
          "suports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuportEntity"
            }
          },
          "favorites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EstablishmentEntity"
            }
          },
          "establishment_suggestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuggestEstablishmentEntity"
            }
          },
          "hashedRefreshToken": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "accessibilities",
          "reviews",
          "suports",
          "favorites",
          "establishment_suggestions",
          "hashedRefreshToken",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "AccessibilityEntity": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserEntity"
              }
            ]
          },
          "establishment": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/EstablishmentEntity"
              }
            ]
          },
          "elevator": {
            "type": "boolean"
          },
          "bar": {
            "type": "boolean"
          },
          "unevenness": {
            "type": "boolean"
          },
          "incompatible_dimensions": {
            "type": "boolean"
          },
          "sign_language": {
            "type": "boolean"
          },
          "tactile_floor": {
            "type": "boolean"
          },
          "braille": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "establishment",
          "elevator",
          "bar",
          "unevenness",
          "incompatible_dimensions",
          "sign_language",
          "tactile_floor",
          "braille",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateRecoverPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "ConfirmTokenDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "token": {
            "type": "number"
          }
        },
        "required": [
          "email",
          "token"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}